feat: Format all the files to pass foundry CI/CD

+++ src/Uniswap.sol
@@ -20,12 +20,12 @@ contract UniswapIntegration {
 
     // Swap exact tokens for another token
     function swapExactTokensForTokens(
-        uint amountIn,
-        uint amountOutMin,
+        uint256 amountIn,
+        uint256 amountOutMin,
         address[] calldata path,
         address to,
-        uint deadline
-    ) external returns (uint[] memory amounts) {
+        uint256 deadline
+    ) external returns (uint256[] memory amounts) {
         require(IERC20(path[0]).transferFrom(msg.sender, address(this), amountIn), "Transfer of token failed");
         IERC20(path[0]).approve(address(uniswapRouter), amountIn);
